fix: move v1.0.0 upgrade helper out of the version directory
@data-fair/lib-node/upgrade-scripts imports every file in
upgrade/<version>/ and reads script.default.description — a helper
module without a default export crashed the worker on boot with
"Cannot read properties of undefined (reading 'description')".

Move plugin-metadata.ts up one level (the framework filters
top-level files out of upgrade/) and update the two import paths.
