+++ upgrade/plugin-metadata.ts
